CourseWork Object

CourseWork Object

gc_course_work(
  materials = NULL,
  associatedWithDeveloper = NULL,
  updateTime = NULL,
  alternateLink = NULL,
  assigneeMode = NULL,
  maxPoints = NULL,
  workType = NULL,
  assignment = NULL,
  multipleChoiceQuestion = NULL,
  description = NULL,
  scheduledTime = NULL,
  creationTime = NULL,
  creatorUserId = NULL,
  individualStudentsOptions = NULL,
  dueDate = NULL,
  submissionModificationMode = NULL,
  state = NULL,
  courseId = NULL,
  id = NULL,
  dueTime = NULL,
  title = NULL,
  topicId = NULL
)

Arguments

materials
Additional materials

associatedWithDeveloper
Whether this course work item is associated with the Developer Console

updateTime
Timestamp of the most recent change to this course work

alternateLink
Absolute link to this course work in the Classroom web UI

assigneeMode
Assignee mode of the coursework

maxPoints
Maximum grade for this course work

workType
Type of this course work

assignment
Assignment details

multipleChoiceQuestion
Multiple choice question details

description
Optional description of this course work

scheduledTime
Optional timestamp when this course work is scheduled to be published

creationTime
Timestamp when this course work was created

creatorUserId
Identifier for the user that created the coursework

individualStudentsOptions
Identifiers of students with access to the coursework

dueDate
Optional date, in UTC, that submissions for this course work are due

submissionModificationMode
Setting to determine when students are allowed to modify submissions

state
Status of this course work

courseId
Identifier of the course

id
Classroom-assigned identifier of this course work, unique per course

dueTime
Optional time of day, in UTC, that submissions for this course work

title
Title of this course work

topicId
Identifier for the topic that this coursework is associated with

Value

CourseWork object

Details

Course work created by a teacher for students of the course.

See also