StudentSubmission Object

StudentSubmission Object

gc_student_submission(
  multipleChoiceSubmission = NULL,
  assignmentSubmission = NULL,
  associatedWithDeveloper = NULL,
  shortAnswerSubmission = NULL,
  updateTime = NULL,
  alternateLink = NULL,
  draftGrade = NULL,
  late = NULL,
  courseWorkType = NULL,
  creationTime = NULL,
  state = NULL,
  userId = NULL,
  courseWorkId = NULL,
  courseId = NULL,
  id = NULL,
  submissionHistory = NULL,
  assignedGrade = NULL
)

Arguments

multipleChoiceSubmission
Submission content when course_work_type is MULTIPLE_CHOICE_QUESTION

assignmentSubmission
Submission content when course_work_type is ASSIGNMENT

associatedWithDeveloper
Whether this student submission is associated with the Developer Console

shortAnswerSubmission
Submission content when course_work_type is SHORT_ANSWER_QUESTION

updateTime
Last update time of this submission

alternateLink
Absolute link to the submission in the Classroom web UI

draftGrade
Optional pending grade

late
Whether this submission is late

courseWorkType
Type of course work this submission is for

creationTime
Creation time of this submission

state
State of this submission

userId
Identifier for the student that owns this submission

courseWorkId
Identifier for the course work this corresponds to

courseId
Identifier of the course

id
Classroom-assigned Identifier for the student submission

submissionHistory
The history of the submission (includes state and grade histories)

assignedGrade
Optional grade

Value

StudentSubmission object

Details

Student submission for course work.StudentSubmission items are generated when a CourseWork item is created.StudentSubmissions that have never been accessed (i.e. with state = NEW)may not have a creation time or update time.

See also