Create Guardian Invitations

Creates a guardian invitation, and sends an email to the guardian askingthem to confirm that they are the student's guardian. Once the guardian accepts the invitation, their state will change toCOMPLETED and they will start receiving guardian notifications. AGuardian resource will also be created to represent the active guardian. The request object must have the student_id andinvited_email_address fields set. Failing to set these fields, orsetting any other fields in the request, will result in an error.

gc_user_profiles_guardian_invitations_create(GuardianInvitation, studentId)

Arguments

GuardianInvitation
The gc_guardian_invitation object to pass to this method

studentId
ID of the student (in standard format)

Details

This method returns the following error codes:

PERMISSION_DENIED if the current user does not have permission to manage guardians, if the guardian in question has already rejected too many requests for that student, if guardians are not enabled for the domain in question, or for other access errors.

RESOURCE_EXHAUSTED if the student or guardian has exceeded the guardian link limit.

INVALID_ARGUMENT if the guardian email address is not valid (for example, if it is too long), or if the format of the student ID provided cannot be recognized (it is not an email address, nor a user_id from this API). This error will also be returned if read-only fields are set, or if the state field is set to to a value other than PENDING.

NOT_FOUND if the student ID provided is a valid student ID, but Classroom has no record of that student.

ALREADY_EXISTS if there is already a pending guardian invitation for the student and invited_email_address provided, or if the provided invited_email_address matches the Google account of an existing Guardian for this user.

Authentication scopes used by this function are:

https://www.googleapis.com/auth/classroom.guardianlinks.students

Set options(googleAuthR.scopes.selected = c("https://www.googleapis.com/auth/classroom.guardianlinks.students") Then run googleAuthR::gar_auth() to authenticate. See gar_auth for details.

See also